move sidechain_network to context_managers
